+++ docs/release_notes.rst
@@ -10,7 +10,9 @@ bpack v1.2.0 (UNRELEASED)
 * Improved documentation and fixed typos.
 * flake8_ configuration moved to a dedicated file.
 * Do not test the :mod:`bpack.ba` backend in PyPy3_.
-* New :func:`bpack.typing.type_params_to_str` function. 
+* New functions:
+  - :func:`bpack.typing.type_params_to_str`
+  - :func:`bpack.descriptors.flat_fields_iterator`
 
 .. _flake8: https://github.com/pycqa/flake8
 .. _PyPy3: https://www.pypy.org